pyproject.toml
[tool.poetry]
name = "tap-testdata"
version = "0.0.1"
description = "`tap-testdata` is a Singer tap for TestData, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Pat Nadolny <firstname.lastname@example.com>"]
keywords = [
    "ELT",
    "TestData",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "<3.12,>=3.8"
singer-sdk = { version="^0.31.1" }
fs-s3fs = { version = "^1.1.1", optional = true }
nodejs-bin = {extras = ["cmd"], version = "^18.4.0a4"}
hypothesis-jsonschema = "^0.22.1"
jsf = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
singer-sdk = { version="^0.31.1", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_testdata"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_testdata"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-testdata = 'tap_testdata.tap:TapTestData.cli'
